mtgWebApp/requirements.txt
certifi==2018.1.18
dj-database-url==0.5.0
Django==2.0.3
django-heroku==0.3.1
django-storages==1.6.6
djangorestframework==3.7.7
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
pipenv==11.8.3
psycopg2==2.7.4
python-dateutil==2.7.2
python-memcached==1.59
pytz==2018.3
s3transfer==0.1.13
six==1.11.0
virtualenv==15.1.0
virtualenv-clone==0.3.0
wfastcgi==3.0.0
whitenoise==3.3.1
pylibmc==1.5.1
